+++ src/plugins/curator/components/gallery-fields/index.jsx
@@ -119,7 +119,7 @@ const GalleryFields = ({ isEditMode, gallery, formName, nodeRef, schemas }) => {
           name="creditUrl"
         />
         {schemas.node.hasMixin('triniti:boost:mixin:sponsorable')
-          && <FieldArray label="Sponsor" name="sponsorRefs" component={SponsorPickerField} isEditMode={isEditMode} />}
+        && <FieldArray label="Sponsor" name="sponsorRefs" component={SponsorPickerField} isEditMode={isEditMode} />}
       </CardBody>
     </Card>
   );