+++ src/plugins/iam/services/Authorizer.js
@@ -1,3 +1,5 @@
+import noop from 'lodash/noop';
+import swal from 'sweetalert2';
 import EventSubscriber from '@gdbots/pbjx/EventSubscriber';
 import { EVENT_PREFIX } from '@gdbots/pbjx/constants';
 import AuthenticationRequired from '../exceptions/AuthenticationRequired';
@@ -29,6 +31,18 @@ export default class Authorizer extends EventSubscriber {
     const state = pbjxEvent.getRedux().getState();
 
     if (!isAuthenticated(state)) {
+      swal.fire({
+        title: 'Session Expired',
+        type: 'error',
+        html: '<p><strong>To avoid losing your work:</strong>' +
+          '<ol class="text-left">' +
+          '<li class="pb-2"><mark><u>DO NOT</u></mark> close or refresh.</li>' +
+          '<li class="pb-2"><a href="/" target="_blank" rel="noopener noreferrer"><strong>Login</strong></a> in a new tab.</li>' +
+          '<li class="pb-2">Once logged in, return to this tab.</li>' +
+          '<li>Click <strong>OK</strong> and then retry your operation.</li>' +
+          '</ol>' +
+          '</p>',
+      }).then(noop).catch(console.error);
       throw new AuthenticationRequired();
     }